agent/requirements.txt
pynput==1.7.6
pygetwindow==0.0.9
requests==2.31.0
psutil==5.9.8
python-dotenv==1.0.1
Pillow==10.2.0